CSV uploading fixed

+++ MeeseeksInc/inventory/api.py
@@ -833,8 +833,9 @@ class ItemUpload(APIView):
     Uploading items via API
     """
     def errorHandling(self, request, message, createdItems):
-        for createdItem in createdItems:
-            createdItem.delete()
+        if createdItems:
+            for createdItem in createdItems:
+                createdItem.delete()
         messages.error(request._request, message)
         return Response(message, status=status.HTTP_400_BAD_REQUEST)
         
@@ -848,38 +849,33 @@ class ItemUpload(APIView):
             if not (header.lower() == "item name" or header.lower() == "quantity" or header.lower() == "model number" or header.lower() =="description" or header.lower() == "tags"):
                 # ERROR CHECK, make sure the custom field names are correct 
                 if not any(field.field_name == header for field in custom_fields):
-                    messages.error(request._request, 
-                                 'field ' + header + ' does not exist')
-                    return Response('field ' + header + ' does not exist', status=status.HTTP_400_BAD_REQUEST) 
+                    return self.errorHandling(request, 'field ' + header + ' does not exist', [])
                 customFieldMap[header.lower()] = i
             else:
                 headerMap[header.lower()] = i    
             
         # ERROR CHECK, make sure that item name and quantity headers exist
         if not "item name" in headerMap:
-            messages.error(request._request, 
-                                 'Item Name does not exist in header')
-            return Response("Make sure item name exists in header", status=status.HTTP_400_BAD_REQUEST) 
+            return self.errorHandling(request, '"Item Name" does not exist in header', [])
         if not "quantity" in headerMap:
-            messages.error(request._request, 
-                                 'Quantity does not exist in header')
-            return Response("Make sure quantity exists in header", status=status.HTTP_400_BAD_REQUEST) 
+            return self.errorHandling(request, '"Quantity" does not exist in header', [])
         
         createdItems = []
         for i, csvRow in enumerate(csvData[1:]):
             row = csvRow.split(',')
             if row[headerMap["item name"]]=='':
-                messages.error(request._request, 
-                                 'item name does not exist in row ' + str(i+1))
-                return Response('item name does not exist in row ' + str(i+1), status=status.HTTP_400_BAD_REQUEST) 
+                return self.errorHandling(request, 'value of "Item Name" does not exist in row ' + str(i+1), createdItems)
             if row[headerMap["quantity"]]=='':
-                messages.error(request._request, 
-                                 'quantity does not exist in row ' + str(i+1))
-                return Response('quantity does not exist in row ' + str(i+1), status=status.HTTP_400_BAD_REQUEST) 
-            item, created = Item.objects.get_or_create(item_name=row[headerMap["item name"]], quantity=row[headerMap["quantity"]], model_number=row[headerMap["model number"]], description=row[headerMap["description"]])
-            if not created:
+                return self.errorHandling(request, 'value of "Quantity" does not exist in row ' + str(i+1), createdItems)
+            if not row[headerMap["quantity"]].isdigit():
+                return self.errorHandling(request, 'value of "Quantity" is not an integer in row' + str(i+1), createdItems)          
+            if int(row[headerMap["quantity"]])<0:
+                return self.errorHandling(request, 'value of "Quantity" is less than 0 in row ' + str(i+1), createdItems)           
+            existingItem = Item.objects.filter(item_name=row[headerMap["item name"]]).count()
+            if existingItem>0:
                 return self.errorHandling(request, "Item " + row[headerMap["item name"]] + " already exists", createdItems)
-
+            item = Item(item_name=row[headerMap["item name"]], quantity=row[headerMap["quantity"]], model_number=row[headerMap["model number"]] if "model number" in headerMap else "", description=row[headerMap["description"]] if "description" in headerMap else "")
+            
             item.save()
             createdItems.append(item)
             # add to tags
@@ -896,13 +892,13 @@ class ItemUpload(APIView):
                         value = Custom_Field_Value(item=item, field=actual_field, field_value_short_text=row[j])
                         value.save()
                     else:
-                        return self.errorHandling(request, "custom_field at row " + str(i+1) + " is not short text. Length is too long", createdItems) 
+                        return self.errorHandling(request, actual_field.field_name + " at row " + str(i+1) + " is not short text. Length is too long", createdItems) 
                 elif actual_field.field_type == "Long":
                     if len(row[j])<=1000:
                         value = Custom_Field_Value(item=item, field=actual_field, field_value_long_text=row[j])
                         value.save()
                     else:
-                        return self.errorHandling(request, "custom_field at row " + str(i+1) + " is not long text. Length is too long", createdItems)  
+                        return self.errorHandling(request, actual_field.field_name + " at row " + str(i+1) + " is not long text. Length is too long", createdItems)  
                 elif actual_field.field_type == "Int":
                     try:
                         int(row[j])
@@ -912,7 +908,7 @@ class ItemUpload(APIView):
                         if row[j] == "":
                             continue
                         else:
-                            return self.errorHandling(request,"custom_field at row " + str(i+1) + " is not a int", createdItems) 
+                            return self.errorHandling(request, actual_field.field_name + " at row " + str(i+1) + " is not an integer", createdItems) 
                 elif actual_field.field_type == "Float":
                     try:
                         float(row[j])
@@ -922,7 +918,7 @@ class ItemUpload(APIView):
                         if row[j] == "":
                             continue
                         else:
-                            return self.errorHandling(request, "custom_field at row " + str(i+1) + " is not a float", createdItems)
+                            return self.errorHandling(request, actual_field.field_name + " at row " + str(i+1) + " is not a float", createdItems)
         items = {}
 #         pk_list = [5, 7, 1, 3, 4]  
 #         clauses = ' '.join(['WHEN id=%s THEN %s' % (pk, i) for i, pk in enumerate(pk_list)])
